Add VersionedApp
