+++ src/main/java/mesosphere/marathon/client/model/v2/QueueElement.java
@@ -3,7 +3,7 @@ package mesosphere.marathon.client.model.v2;
 public class QueueElement {
     private int count;
     private Delay delay;
-    private App app;
+    private VersionedApp app;
 
     public int getCount() {
         return count;
@@ -21,11 +21,11 @@ public class QueueElement {
         this.delay = delay;
     }
 
-    public App getApp() {
+    public VersionedApp getApp() {
         return app;
     }
 
-    public void setApp(App app) {
+    public void setApp(VersionedApp app) {
         this.app = app;
     }
 }
